extract concerns
